marketplace/quant/kozmoai_platform/core/pyproject.toml
[tool.poetry]
name = "kozmoai-core"
version = "1.4.2"
description = "KozmoAI package with core functionality."
authors = ["KozmoAI Team <hello@kozmoai.co>"]
license = "AGPL-3.0-only"
readme = "README.md"
packages = [{ include = "kozmoai_core" }]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.34.0"
websockets = "^14.0"
pandas = ">=1.5.3"
html5lib = "^1.1"
fastapi = "^0.115"
uuid7 = "^0.1.0"
posthog = "^3.3.1"
python-multipart = "^0.0.18"
pydantic = "^2.5.1"
requests = "^2.32.1"
importlib-metadata = ">=6.8.0"
python-dotenv = "^1.0.0"
aiohttp = "^3.11.11"
ruff = "^0.7"              # Needed here to lint generated code
pyjwt = "^2.10.1"

[tool.poetry.scripts]
kozmoai-build = "kozmoai_core.build:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
